Feature/group by iduser

+++ src/controllers/group/index.ts
@@ -56,12 +56,10 @@ router.delete('/group/delete/:id/:userId', async (req: Request, res: Response) =
     if (!group) {
       return res.status(404).json({ error: 'Grupo não encontrado' });
     }
-    // const userCreated = await Group.findById({ 'createdByUser.user._id': userId })
-    // // const userExists = group.createdByUser.find((createdByUser) => createdByUser.user._id.toString() === userId.toString());
-    // console.log(userCreated)
-    // if (!userCreated) {
-    //   return res.status(401).json({ error: 'Você não tem permissão para excluir este grupo' });
-    // }
+    const userCreated = await Group.findOne({ 'createdByUser.user._id': userId })
+    if (!userCreated) {
+      return res.status(401).json({ error: 'Você não tem permissão para excluir este grupo' });
+    }
 
     await Group.findByIdAndDelete(id);
     await Users.updateMany({ $pull: { groupIds: id } });
@@ -135,8 +133,9 @@ router.post('/group/add-member/:id/:userId', async (req: Request, res: Response)
     }
     const userExists = await Group.findOne({ 'members.user._id': userId });
     // Verifica se o userId já está presente no array memberIds
+    console.log(userExists)
     if (userExists) {
-      return res.status(400).json({ message: 'Usuário já é membro deste grupo' });
+      return res.status(400).json({ error: 'Usuário já é membro deste grupo' });
     }
 
     await Group.findByIdAndUpdate(id, { $addToSet: { members: { user: user } } });
@@ -246,6 +245,31 @@ router.get('/group/trending-groups', async (req: Request, res: Response) => {
       { $skip: (page - 1) * pageSize }, // Ignorar documentos nas páginas anteriores
       { $limit: pageSize } // Limitar o número de documentos retornados por página
     ]);
+    res.status(200).json({
+      groups,
+      page,
+      pageSize,
+      totalCount,
+      totalPages
+    });
+  } catch (error) {
+    console.error('Error:', error);
+    res.status(500).json({ error: error });
+  }
+});
+router.get('/group/get-by-user-id/:userId', async (req: Request, res: Response) => {
+  try {
+    const userId: string = req.params.userId;
+    const page = parseInt(req.query.page as string) || 1;
+    const pageSize = parseInt(req.query.pageSize as string) || 10;
+    if (!userId) {
+      return res.status(422).json({ error: 'Informe um usuário!' });
+    }
+    const totalCount = await Group.countDocuments(); // Total de documentos 
+    const totalPages = Math.ceil(totalCount / pageSize); // Total de páginas
+    const groups = await Group.find({ 'members.user._id': userId })
+      .skip((page - 1) * pageSize) // Ignorar documentos nas páginas anteriores
+      .limit(pageSize); // Limitar o número de documentos retornados por página
 
     res.status(200).json({
       groups,
